requirements.txt
streamlit
langchain-groq
python-dotenv
chunking-evaluation @ https://github.com/brandonstarxel/chunking_evaluation.git